- Added MessageData type - Added HighRes Timer to the Engine
Added a MessageData type so that the Message type can be left alone regardless of type of message and the MessageData type can be inherited from for different use cases. However, its likely that there will be a generic GameMessageData type with a bunch of vectors, and floats, and such like that will probably suffice for most use cases.

The Timer is meant for getting measures of frameNumber, fps, delta frame time, etc.
Also, the Timer will raise FRAME events for objects to subscribe to.